moves scValByType to helpers, adds tests for scValByType address types
